requirements.txt
Flask>=2.0
python-dotenv>=0.19
Flask-Cors>=3.0
python-magic>=0.4
Pillow>=9.0
requests>=2.25
openai>=1.0 # Added for OpenAI API access
google-generativeai>=0.3.0 # Added for Google Gemini API access
